# packages/contracts/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AEP BFF API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /projects/{projectName}/agents/{conversationId}/messages:
    get:
      operationId: get-conversation
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: FE-chosen conversation uuid
        in: path
        name: conversationId
        required: true
        schema:
          description: FE-chosen conversation uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Rehydrate a chat conversation (messages only)
      tags:
      - GenAI
    post:
      operationId: create-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: FE-chosen conversation uuid
        in: path
        name: conversationId
        required: true
        schema:
          description: FE-chosen conversation uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./components.yaml#/components/schemas/TurnInputBody
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/TurnOutputBody
          description: Accepted
        '409':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/TurnConflict
          description: A turn is already running for this project, or the design
            use-case has no requirements yet.
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Start a generation/chat turn (202 — runs detached; attach via the turn stream)
      tags:
      - GenAI
...
  /projects/{projectName}/turns/active:
    get:
      operationId: get-active-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/TurnStatus
          description: OK
        '204':
          description: No running turn
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Get the project's running turn (204 when none)
      tags:
      - GenAI
  /projects/{projectName}/turns/{turnId}:
    get:
      operationId: get-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Turn id from the 202 create response
        in: path
        name: turnId
        required: true
        schema:
          description: Turn id from the 202 create response
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/TurnStatus
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Get one turn's status
      tags:
      - GenAI
  /projects/{projectName}/turns/{turnId}/stream:
    get:
      operationId: stream-turn
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Turn id from the 202 create response
        in: path
        name: turnId
        required: true
        schema:
          description: Turn id from the 202 create response
          type: string
      - description: Replay from this absolute event index (wins over Last-Event-ID)
        in: query
        name: from
        schema:
          description: Replay from this absolute event index (wins over Last-Event-ID)
          minimum: -1
          type: integer
      - description: 'SSE auto-reconnect resume cursor: the last frame id the client saw; replay
          resumes after it. Opaque per the SSE spec — non-numeric values are ignored (full replay).
          The from query param wins when both are present.'
        in: header
        name: Last-Event-ID
        schema:
          type: string
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                description: Raw agent StreamPart frames + one terminal event, id-stamped; ends with [DONE]
                type: string
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Attach to a turn's event stream (SSE replay + live tail)
      tags:
      - GenAI
...
components:
  securitySchemes:
    userJWT:
      bearerFormat: JWT
      description: Thunder end-user JWT (RS256, JWKS-verified)
      scheme: bearer
      type: http